Add SFX as simple embedded ogg vorbis files
