+++ Shared/SFX/SFXLoader.cs
@@ -1,0 +1,377 @@
+﻿using KK_VR.Interpreters;
+using System;
+using System.Collections;
+using System.Collections.Generic;
+using System.IO;
+using UnityEngine.Networking;
+using UnityEngine;
+using VRGIN.Core;
+using System.Reflection;
+
+namespace KK_VR.Holders
+{
+    /// <summary>
+    /// Provides SFX for character <-> controller interactions
+    /// </summary>
+    internal class SFXLoader
+    {
+        // Sfx contributions are greatly appreciated.
+        // Mix them in Audacity or something.
+
+        private readonly AudioSource _audioSource;
+        internal bool IsPlaying => _audioSource.isPlaying;
+
+        private static readonly Dictionary<Sfx, List<List<List<AudioClip>>>> sfxDic = [];
+        internal SFXLoader(AudioSource audioSource)
+        {
+            _audioSource = audioSource;
+        }
+        internal static void Init()
+        {
+            InitDic();
+            LoadEmbeddedResources();
+        }
+        internal void PlaySfx(float volume, Sfx sfx, Surface surface, Intensity intensity, bool overwrite)
+        {
+            if (!KoikatuInterpreter.Settings.EnableSFX) return;
+            if (_audioSource.isPlaying)
+            {
+                if (!overwrite) return;
+                _audioSource.Stop();
+            }
+
+            //VRPlugin.Logger.LogInfo($"AttemptToPlay:{sfx}:{surface}:{intensity}:{volume}");
+            AdjustInput(sfx, ref surface, ref intensity);
+            var audioClipList = sfxDic[sfx][(int)surface][(int)intensity];
+            var count = audioClipList.Count;
+            if (count != 0)
+            {
+                _audioSource.volume = Mathf.Clamp01(volume);
+                _audioSource.pitch = 0.9f + UnityEngine.Random.value * 0.2f;
+                _audioSource.clip = audioClipList[UnityEngine.Random.Range(0, count)];
+                _audioSource.Play();
+            }
+
+        }
+        private void AdjustInput(Sfx sfx, ref Surface surface, ref Intensity intensity)
+        {
+            // Because currently we have far from every category covered.
+            if (intensity == Intensity.Wet)
+            {
+                surface = Surface.Skin;
+            }
+            else if (sfx == Sfx.Slap)
+            {
+                surface = Surface.Skin;
+            }
+            else if (sfx == Sfx.Traverse)
+            {
+                if (surface == Surface.Hair)
+                {
+                    intensity = Intensity.Soft;
+                }
+                else if (surface == Surface.Skin)
+                {
+                    intensity = Intensity.Soft;
+                }
+            }
+            else if (sfx == Sfx.Undress)
+            {
+
+            }
+        }
+
+        // As of now categories are highly inconsistent, perhaps revamp of sorts?
+        internal enum Sfx
+        {
+            Tap,
+            Slap,
+            Traverse,
+            Undress,
+        }
+
+        internal enum Surface
+        {
+            Skin,
+            Cloth,
+            Hair
+        }
+
+        internal enum Intensity
+        {
+            // Think about:
+            //     Soft as something smallish and soft and on slower side of things, like boobs or ass.
+            //     Rough as something flattish and big and at times intense, like tummy or thighs.
+            //     Wet as.. yet to mix something proper for it. WIP.
+            Soft,
+            Rough,
+            Wet
+        }
+
+        private static void InitDic()
+        {
+            var sfxNames = Enum.GetNames(typeof(Sfx));
+            var surfaceNames = Enum.GetNames(typeof(Surface));
+            var intenseNames = Enum.GetNames(typeof(Intensity));
+
+            for (var i = 0; i < sfxNames.Length; i++)
+            {
+                var key = (Sfx)i;
+                sfxDic.Add(key, []);
+                for (var j = 0; j < surfaceNames.Length; j++)
+                {
+                    sfxDic[key].Add([]);
+                    for (var k = 0; k < intenseNames.Length; k++)
+                    {
+                        sfxDic[key][j].Add([]);
+                    }
+                }
+            }
+        }
+
+        private static bool FindIndex(string[] array, string name, out int index)
+        {
+            index = Array.FindIndex(array, n => n.Equals(name, StringComparison.OrdinalIgnoreCase));
+            return index != -1;
+        }
+
+#if KK
+        // .Net Framework 3.5 lacks this method.
+        public static void CopyStream(Stream source, Stream destination, int bufferSize = 81920)
+        {
+            byte[] buffer = new byte[bufferSize];
+            int bytesRead;
+
+            while ((bytesRead = source.Read(buffer, 0, buffer.Length)) > 0)
+            {
+                destination.Write(buffer, 0, bytesRead);
+            }
+        }
+#endif
+
+        private static void LoadEmbeddedResources()
+        {
+            var assembly = Assembly.GetAssembly(typeof(SFXLoader));
+            var resources = assembly.GetManifestResourceNames();
+
+            var sfxNames = Enum.GetNames(typeof(Sfx));
+            var surfaceNames = Enum.GetNames(typeof(Surface));
+            var intenseNames = Enum.GetNames(typeof(Intensity));
+
+            foreach (var resource in resources)
+            {
+                if (!resource.EndsWith(".ogg", StringComparison.OrdinalIgnoreCase)) continue;
+
+                var folders = resource.Trim().Split('.');
+
+                if (folders.Length > 6)
+                {
+                    // Approx estimated structure:
+                    // KKS_VR.SFX.Slap.Skin.Rough.sv_se_h_44.ogg
+
+                    // Split into:
+                    // KKS_VR   SFX   Slap   Skin   Rough   sv_se_h_44   ogg
+
+                    if (FindIndex(sfxNames, folders[2], out var sfxIndex)
+                        && FindIndex(surfaceNames, folders[3], out var surfaceIndex)
+                        && FindIndex(intenseNames, folders[4], out var intenseIndex))
+                    {
+
+
+                        // Temp path + file name + extension
+                        var tempPath = System.IO.Path.Combine(
+                            System.IO.Path.GetTempPath(),  
+                            folders[folders.Length - 2] + "." + folders[folders.Length - 1]
+                            );
+
+                        using (var stream = assembly.GetManifestResourceStream(resource))
+                        {
+                            using var fileStream = new FileStream(tempPath, FileMode.Create, FileAccess.ReadWrite);
+#if KK
+                            CopyStream(stream, fileStream);
+#else
+                            stream.CopyTo(fileStream);
+#endif
+                            // Should be automatic
+                            //stream.Close();
+                            //fileStream.Close();
+                        }
+
+                        // Add to list on finished loading
+                        var list = sfxDic[(Sfx)sfxIndex][surfaceIndex][intenseIndex];
+
+                        // Some component's instance to run the coroutine
+                        VRManager.Instance.StartCoroutine(
+                            LoadAudioFile(
+                                BepInEx.Utility.ConvertToWWWFormat(tempPath), 
+                                folders[folders.Length -2], 
+                                list)
+                            );
+                    }
+                }
+            }
+        }
+
+        private static IEnumerator LoadAudioFile(string path, string name, List<AudioClip> destination)
+        {
+
+#if KK
+            var audioFile = UnityWebRequest.GetAudioClip(path, AudioType.OGGVORBIS);
+            
+            yield return audioFile.Send();
+            if (audioFile.isError)
+#else
+            var audioFile = UnityWebRequestMultimedia.GetAudioClip(path, AudioType.OGGVORBIS);
+
+            yield return audioFile.SendWebRequest();
+            if (audioFile.isHttpError || audioFile.isNetworkError)
+#endif
+            {
+                VRPlugin.Logger.LogWarning($"{audioFile.error} - {path}");
+            }
+            else
+            {
+                var clip = DownloadHandlerAudioClip.GetContent(audioFile);
+                clip.name = name;
+                destination.Add(clip);
+#if DEBUG
+                VRPlugin.Logger.LogDebug($"Loaded:SFX:{path}");
+#endif
+            }
+        }
+
+        //        private static void Populate()
+        //        {
+        //            var assembly = Assembly.GetAssembly(typeof(SFXLoader));
+        //            var resources = assembly.GetManifestResourceNames();
+        //            var sfxNames = Enum.GetNames(typeof(Sfx));
+        //            var surfaceNames = Enum.GetNames(typeof(Surface));
+        //            var intenseNames = Enum.GetNames (typeof(Intensity));
+        //            foreach (var resource in resources)
+        //            {
+        //                var stream = assembly.GetManifestResourceStream(resource);
+        //                if (stream == null)
+        //                {
+        //                    VRPlugin.Logger.LogWarning($"Failed to get stream from resource\n{resource}");
+        //                    continue;
+        //                }
+
+        //                var assetBundle =
+        //#if KK
+        //                    AssetBundle.LoadFromMemory(ResourceUtils.ReadAllBytes(stream));
+        //#else
+        //                    AssetBundle.LoadFromStream(stream);
+        //#endif
+        //                stream.Close();
+
+        //                if (assetBundle == null)
+        //                {
+        //                    VRPlugin.Logger.LogWarning($"Failed to load embedded resource\n{resource}");
+        //                    continue;
+        //                }
+
+        //                foreach (var asset in assetBundle.GetAllAssetNames())
+        //                {
+        //                    var words = asset.Trim().Split('/');
+        //                    if (words.Length > 5
+        //                        && words[0].Equals("assets", StringComparison.OrdinalIgnoreCase)
+        //                        && words[1].Equals("sfx", StringComparison.OrdinalIgnoreCase))
+        //                    {
+        //                        if (FindIndex(sfxNames, words[2], out var sfxIndex)
+        //                            && FindIndex(surfaceNames, words[3], out var surfaceIndex)
+        //                            && FindIndex(intenseNames, words[4], out var intenseIndex))
+        //                        {
+        //                            var audioClip = assetBundle.LoadAsset<AudioClip>(words[5]);
+        //                            if (audioClip != null)
+        //                            {
+        //#if DEBUG
+        //                                VRPlugin.Logger.LogDebug($"AddEmbeddedAsset - {words[5]} : {asset}");
+        //#endif
+        //                                sfxDic[(Sfx)sfxIndex][surfaceIndex][intenseIndex].Add(audioClip);
+        //                            }
+        //                        }
+        //                    }
+        //                }
+        //            }
+        //        }
+        //private static void PopulateDic()
+        //{
+
+        //    //Populate();
+        //    //return;
+        //    //for (var i = 0; i < sfxDic.Count; i++)
+        //    //{
+        //    //    var key = (Sfx)i;
+        //    //    for (var j = 0; j < sfxDic[key].Count; j++)
+        //    //    {
+        //    //        for (var k = 0; k < sfxDic[key][j].Count; k++)
+        //    //        {
+        //    //            //var directory = BepInEx.Utility.CombinePaths(
+        //    //            //    [
+        //    //            //        Paths.PluginPath,
+        //    //            //        "SFX",
+        //    //            //        key.ToString(),
+        //    //            //        ((Surface)j).ToString(),
+        //    //            //        ((Intensity)k).ToString()
+        //    //            //    ]);
+        //    //            //if (Directory.Exists(directory))
+        //    //            //{
+
+        //    //                var dirInfo = new DirectoryInfo(directory);
+        //    //                var clipNames = new List<string>();
+        //    //                //foreach (var file in dirInfo.GetFiles("*.wav"))
+        //    //                //{
+        //    //                //    clipNames.Add(file.Name);
+        //    //                //}
+        //    //                foreach (var file in dirInfo.GetFiles("*.ogg"))
+        //    //                {
+        //    //                    clipNames.Add(file.Name);
+        //    //                }
+        //    //                if (clipNames.Count == 0) continue;
+        //    //                VRManager.Instance.StartCoroutine(LoadAudioFile(directory, clipNames, sfxDic[key][j][k]));
+        //    //            //}
+        //    //        }
+        //    //    }
+        //    //}
+        //}
+
+        //        private static IEnumerator LoadAudioFile(string path, List<string> clipNames, List<AudioClip> destination)
+        //        {
+        //            foreach (var name in clipNames)
+        //            {
+        //                //UnityWebRequest audioFile;
+        //                //if (name.EndsWith(".wav"))
+        //                //{
+        //                //    audioFile = UnityWebRequest.GetAudioClip(Path.Combine(path, name), AudioType.WAV);
+        //                //}
+        //                //else
+        //                //{
+        //#if KK
+
+        //                var audioFile = UnityWebRequest.GetAudioClip(System.IO.Path.Combine(path, name), AudioType.OGGVORBIS);
+        //#else
+        //                var audioFile = UnityWebRequestMultimedia.GetAudioClip(System.IO.Path.Combine(path, name), AudioType.OGGVORBIS);
+        //#endif
+        //                //}
+        //#if KK
+
+        //                yield return audioFile.Send();
+        //                if (audioFile.isError)
+        //#else
+        //                yield return audioFile.SendWebRequest();
+        //                if (audioFile.isHttpError || audioFile.isNetworkError)
+        //#endif
+        //                {
+        //                    VRPlugin.Logger.LogWarning($"{audioFile.error} - {System.IO.Path.Combine(path, name)}");
+        //                }
+        //                else
+        //                {
+        //                    var clip = DownloadHandlerAudioClip.GetContent(audioFile);
+        //                    clip.name = name;
+        //                    destination.Add(clip);
+        //                    //VRPlugin.Logger.LogDebug($"Loaded:SFX:{name}");
+        //                }
+        //            }
+        //        }
+    }
+}
