+++ Shared/SFX/Slap/Skin/Rough/sv_se_h_42.ogg
